finish

+++ app/db/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, String, Float, ForeignKey, DateTime, Numeric
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from .base import Base
@@ -9,7 +9,7 @@ class Lot(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     title = Column(String, nullable=False)
-    start_price = Column(Float, nullable=False)
+    start_price = Column(Numeric(10, 2), nullable=False)
     current_price = Column(Float, nullable=False)
     status = Column(String, default="running")
     end_time = Column(DateTime)
